Criação de arquivo para gerenciamento de fases
